# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API",
    "version": "1.0.0",
    "description": "A secure API demonstrating Supabase authentication with protected routes."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/public/info": {
      "get": {
        "summary": "Public info, no auth required",
        "responses": {
          "200": { "description": "Public welcome message" }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "you@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing fields or invalid signup" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "you@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns access_token" },
          "400": { "description": "Missing fields" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the logged-in user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile data" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Access the protected dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard welcome message" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    }
  }
}